refactor(playground): move agent workflow components
