+++ components/playground/ContainerEnvironmentManager.tsx
